requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.1.7
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
gunicorn==20.1.0
idna==3.4
isort==5.10.1
mccabe==0.7.0
pep8-naming==0.13.2
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
requests==2.28.1
schedule==1.1.0
#scheduler==0.8.1
sqlparse==0.4.3
typeguard==2.13.3
tzdata==2022.6
urllib3==1.26.12
